agentic-malloy: layer-improve review fixes (P1 train-only, P2 source-scoped views, P3 split)
Addresses review findings on the merged layer-improve work.

P1 (train-only edits): a layer edit (and skill-fix application) is now REFUSED
when the --from run includes any held-out/test task — checked against
data/split.json train_ids, not the row's recorded `split`. Previously any run,
including test results, could tune the layer and keep model_authored → pass the
official gate. Held-out runs are now triaged/reported only; provenance lineage
records from_splits/train_only.

P2 (source-scoped views): Malloy views are source-scoped and this layer reuses
view names across sources (by_account_type, by_mcc, listing), so the old global
view→file map probed/edited the wrong file. The index is now source-keyed
(source names ARE globally unique; a view lives in its source's file), and
referencedViews resolves against the query's HEAD SOURCE via parseHeadSource.

P3 (module split): the 1134-line module is split into focused units —
miss-analysis (parse/index/classify/evidence), miss-verdict (model manner+owner
+fix, tool diagnosis, trace block), layer-repair (edit/gate/snapshot/restore/
provenance), and a slim layer-improve orchestrator. run-log unchanged.

Verified: npx tsc --noEmit clean; 80/80 vitest (+ P2 duplicate-view resolution,
P1 nonTrainTaskIds, parseHeadSource). Live: a synthetic held-out miss is refused
($0); the real layer's duplicate views resolve source-scoped.
