+++ projects/agentic-malloy/src/layer-improve.test.ts
@@ -1,11 +1,11 @@
 /**
- * layer-improve unit tests. Two pieces the handoff calls out specifically:
- *  - the MISS → FILE mapping (which layer file a submitted query implicates), and
- *  - the LAYER-vs-SKILL triage (classifyMiss, a pure function over re-execution
- *    evidence — the hard, valuable part).
- * Plus a no-leakage assertion: the evidence fed to the model never carries the
- * gold answer. No credentials / runtime needed — the structural-defect probe is
- * modeled as plain MissAnalysis inputs.
+ * layer-improve unit tests:
+ *  - the MISS → FILE mapping, resolved via SOURCE names (P2: views are
+ *    source-scoped and reused across sources, so a global view→file map is wrong);
+ *  - the LAYER-vs-SKILL triage (classifyMiss, pure over re-execution evidence);
+ *  - the TRAIN-ONLY guard (P1: held-out task ids must not drive layer edits);
+ *  - no-leakage: the evidence/trace fed to the model never carries the gold answer.
+ * No credentials/runtime needed — the structural probe is modeled as plain inputs.
  */
 import { describe, it, expect } from 'vitest';
 import {
@@ -13,22 +13,32 @@ import {
   mapMalloyToFiles,
   referencedViews,
   identifiersIn,
+  parseHeadSource,
   classifyMiss,
   evidenceBlock,
-  traceBlock,
   analyzeMiss,
+  nonTrainTaskIds,
   type MissAnalysis,
   type MissClassification,
-} from './layer-improve.js';
+} from './miss-analysis.js';
+import { traceBlock } from './miss-verdict.js';
 import type { TaskTrace } from './run-log.js';
 import type { MalloyRuntime } from './malloy-runtime.js';
 
-// A miniature layer mirroring the real f2163373 structure: a lens source in
-// dabstep.malloy, a fee-match source + views in c3, base views in fees_base.
+// A miniature layer mirroring the real f2163373 structure — and crucially the
+// DUPLICATE view name `by_account_type` on two different sources/files (as the
+// real layer has on payments_enriched + merchants_base), which a global view
+// index would mis-resolve.
 const BODIES: Record<string, string> = {
   'fees_base.malloy': `source: fees_base is duckdb.table('fees') extend {
   measure: avg_fee is avg(fee)
   view: by_card_scheme is { group_by: card_scheme; aggregate: avg_fee }
+}`,
+  'merchants_base.malloy': `source: merchants_base is duckdb.table('merchants') extend {
+  view: by_account_type is { group_by: account_type; aggregate: n is count() }
+}`,
+  'c1_payments_enriched.malloy': `source: payments_enriched is payments_base extend {
+  view: by_account_type is { group_by: acct; aggregate: total is sum(amount) }
 }`,
   'dabstep.malloy': `source: rules_lens is fees_base extend {
   dimension: avg_fee_1000eur is fixed_amount + rate
@@ -43,38 +53,57 @@ const BODIES: Record<string, string> = {
 
 const INDEX = buildLayerIndex(BODIES);
 
-describe('miss → file mapping', () => {
-  it('indexes every source and view to its defining file', () => {
-    expect(INDEX.sources.has('fees_base')).toBe(true);
+describe('miss → file mapping (source-keyed)', () => {
+  it('indexes each source to its file and each source to its own views', () => {
     expect(INDEX.sources.has('rules_lens')).toBe(true);
     expect(INDEX.sources.has('fee_match')).toBe(true);
-    expect(INDEX.views.has('total_fees_by_merchant_year')).toBe(true);
-    expect(INDEX.fileOf.get('rules_lens')).toBe('dabstep.malloy');
-    expect(INDEX.fileOf.get('total_fees_by_merchant_year')).toBe('c3_fee_assignment.malloy');
-    expect(INDEX.fileOf.get('by_card_scheme')).toBe('fees_base.malloy');
+    expect(INDEX.fileOfSource.get('rules_lens')).toBe('dabstep.malloy');
+    expect(INDEX.fileOfSource.get('fee_match')).toBe('c3_fee_assignment.malloy');
+    expect(INDEX.viewsBySource.get('fee_match')!.has('total_fees_by_merchant_year')).toBe(true);
+    expect(INDEX.viewsBySource.get('fees_base')!.has('by_card_scheme')).toBe(true);
   });
 
-  it('maps a query on a lens source to the file that defines the source', () => {
+  it('maps a query to the file that defines its source', () => {
     const src = `run: rules_lens -> { where: card_scheme = 'NexPay'; aggregate: m is avg(avg_fee_1000eur) }`;
     expect(mapMalloyToFiles(src, INDEX)).toEqual(['dabstep.malloy']);
   });
 
-  it('maps a bare named-view query to the file defining the source AND the view', () => {
+  it('maps a bare named-view query via the source (view lives in the source file)', () => {
     const src = `run: fee_match -> total_fees_by_merchant_year`;
     expect(mapMalloyToFiles(src, INDEX)).toEqual(['c3_fee_assignment.malloy']);
-    expect(referencedViews(src, INDEX)).toEqual(['total_fees_by_merchant_year']);
+    expect(referencedViews(src, INDEX)).toEqual([{ source: 'fee_match', view: 'total_fees_by_merchant_year' }]);
   });
 
-  it('ignores tokens that are not known sources/views (keywords, fields, literals)', () => {
-    const src = `run: rules_lens -> { group_by: card_scheme; aggregate: avg_fee }`;
-    // card_scheme / avg_fee / group_by are fields/keywords, not source/view names.
-    expect(mapMalloyToFiles(src, INDEX)).toEqual(['dabstep.malloy']);
-    expect(referencedViews(src, INDEX)).toEqual([]);
+  it('P2: a view name shared by two sources resolves to the HEAD source, not a global first-wins', () => {
+    expect(parseHeadSource('run: merchants_base -> by_account_type')).toBe('merchants_base');
+    // Same view name, different head source → different source/file each time.
+    expect(referencedViews('run: merchants_base -> by_account_type', INDEX)).toEqual([{ source: 'merchants_base', view: 'by_account_type' }]);
+    expect(mapMalloyToFiles('run: merchants_base -> by_account_type', INDEX)).toEqual(['merchants_base.malloy']);
+    expect(referencedViews('run: payments_enriched -> by_account_type', INDEX)).toEqual([{ source: 'payments_enriched', view: 'by_account_type' }]);
+    expect(mapMalloyToFiles('run: payments_enriched -> by_account_type', INDEX)).toEqual(['c1_payments_enriched.malloy']);
   });
 
-  it('returns no files for an empty / non-submission', () => {
+  it('does not attribute a view to a source that does not define it', () => {
+    // by_account_type is NOT a view of rules_lens, so referencing it under
+    // rules_lens yields no probe (avoids probing the wrong source).
+    expect(referencedViews('run: rules_lens -> by_account_type', INDEX)).toEqual([]);
+  });
+
+  it('ignores non-source/view tokens and handles empty input', () => {
+    expect(mapMalloyToFiles('run: rules_lens -> { group_by: card_scheme; aggregate: avg_fee }', INDEX)).toEqual(['dabstep.malloy']);
     expect(mapMalloyToFiles('', INDEX)).toEqual([]);
     expect(identifiersIn('a a b a')).toEqual(['a', 'b']);
+  });
+});
+
+// --- P1: train-only guard ----------------------------------------------------
+
+describe('train-only guard (nonTrainTaskIds)', () => {
+  const train = new Set(['1290', '1451', '1507']);
+  it('flags task ids outside the train split (held-out → no layer edits)', () => {
+    expect(nonTrainTaskIds(['1290', '1451'], train)).toEqual([]); // all train
+    expect(nonTrainTaskIds(['1290', '9999'], train)).toEqual(['9999']); // held-out present
+    expect(nonTrainTaskIds([1290, 9999], train)).toEqual(['9999']); // numeric ids coerced
   });
 });
 
@@ -102,8 +131,6 @@ describe('layer-vs-skill triage (classifyMiss)', () => {
   });
 
   it('1290-style: query re-runs fine and returns rows → SKILL (agent inline logic), NOT layer', () => {
-    // The agent filtered `is_credit = true` and dropped the wildcard rules: the
-    // query is valid and returns a row, just the wrong one. The layer is fine.
     const c = classifyMiss(baseAnalysis({ reExec: { ok: true, rowCount: 1 }, viewProbes: [] }));
     expect(c.category).toBe('query_wrong_answer');
     expect(c.suggestedOwner).toBe('skill');
@@ -120,7 +147,6 @@ describe('layer-vs-skill triage (classifyMiss)', () => {
     expect(c.category).toBe('layer_view_error');
     expect(c.suggestedOwner).toBe('layer');
     expect(c.layerSuspected).toBe(true);
-    // the broken view's file is surfaced first.
     expect(c.implicatedFiles[0]).toBe('c3_fee_assignment.malloy');
   });
 
@@ -165,16 +191,11 @@ describe('layer-vs-skill triage (classifyMiss)', () => {
 
 describe('no leakage', () => {
   it('the evidence block fed to the model never contains the gold answer', async () => {
-    // A fake runtime so analyzeMiss runs without credentials. The row carries a
-    // gold_answer — analyzeMiss must never surface it into the evidence.
-    const fakeRt = {
-      run: async () => ({ ok: true, rows: [{ x: 1 }] }),
-    } as unknown as MalloyRuntime;
+    const fakeRt = { run: async () => ({ ok: true, rows: [{ x: 1 }] }) } as unknown as MalloyRuntime;
     const a = await analyzeMiss(
       { task_id: '1290', question: 'avg fee for NexPay credit?', is_correct: false, submitted: true, malloy_source: 'run: rules_lens -> { aggregate: m is avg(avg_fee_1000eur) }', predicted_answer: '5.757053', gold_answer: '5.715872' } as never,
       fakeRt,
       INDEX,
-      new Map(),
     );
     const cls: MissClassification = classifyMiss(a);
     const ev = evidenceBlock(a, cls);
@@ -199,9 +220,7 @@ describe('no leakage', () => {
     };
     const tb = traceBlock(trace, '5.757053', true);
     expect(tb).toContain('list_malloy_files');
-    expect(tb).toContain('run_malloy');
     expect(tb).toContain('Explored the layer');
-    expect(tb).toContain('Reused a NAMED layer view');
     expect(tb).toContain('scalar'); // answer-shape signal (shape only, not the value)
     expect(tb).not.toContain('5.715872'); // gold never enters the trace evidence
   });
